The code generated page fault during execution of movdqa instruction This SSE instruction requered the data to be alined by 16 bytes.
The code corresponding to the bug is:

@SmalltalkVM::newOrdinaryObject
    for (uint32_t index = 0; index < fieldsCount; index++)
        instance->putField(index, globals.nilObject);

movdqa moves a double quadword from src to dst. The compiler optimized this loop into a set of movdqa instructions.
That means that 'instance' was aligned neither by 16 bytes nor by 4 bytes.

When BakerMemoryManager::growHeap does its job it divides newHeapSize by 2.
 newHeapSize/2 must be equal to correctPadding(newHeapSize/2). But it is not.

Issue: #28